Acemoglou, Johnson, and Robinson (2001) Dataset on Colonial Origins of Comparative Development

Colonial factors influencing economic development. Updated in 2012, it includes variables related to settler mortality, economic indicators, geographic data, and other socio-political metrics.

ajr2001

Format

A data frame with 64 observations across 31 columns. Missing values are coded as NA.

longname

Full country name

shortnam

3-letter country name

logmort0

Original log settler mortality

risk

Expected risk

loggdp

Log GDP per capita (PPP) in 1995

campaign

Original mortality from campaign

source0

Country source of own mortality rate

slave

Data from forced labor

latitude

Absolute value of latitude

neoeuro

Indicator for 'New-Europes' (USA, CAN, AUS, NZL)

asia

Indicator for Asia

africa

Indicator for Africa

other

Indicator for 'Other' continent (AUS, NZL, MLT)

edes1975

Population of European descent, 1975

campaignsj

Indicator for HKG and NZL

campaignsj2

Extended recoding of campaign variable

mortnaval1

Mortality using naval stations, method 1

logmortnaval1

Log mortality using naval stations, method 1

mortnaval2

Mortality using naval stations, method 2

logmortnaval2

Log mortality using naval stations, method 2

mortjam

Mortality benchmarked to the Caribbean

logmortjam

Log mortality benchmarked to the Caribbean

logmortcap250

Log mortality capped at 250

logmortjam250

Log mortality benchmarked to the Caribbean, capped at 250

wandcafrica

Indicator for West and Central Africa

malfal94

Falciparum malaria index, 1994

wacacontested

Indicator for contested observations in West and Central Africa

mortnaval2250

Mortality using naval stations, capped at 250

logmortnaval2250

Log mortality using naval stations, capped at 250

mortnaval1250

Mortality using naval stations, capped at 250

logmortnaval1250

Log mortality using naval stations, capped at 250

Source

Acemoglou, D., Johnson, S., & Robinson, J. A. (2001). "The Colonial Origins of Comparative Development: An Empirical Investigation." American Economic Review. Updated version (2012): "The Colonial Origins of Comparative Development: An Empirical Investigation: Reply." Data available at https://www.aeaweb.org/aer/data/oct2012/20110390_data.zip