Bernheim, Meer, and Novarro (2016) Dataset on Liquor Consumption and Commitment Opportunities

Liquor sale hours, alcohol consumption, and related state policies across 47 U.S. states from 1970 to 2007. The dataset includes off-premises and on-premises sales data for beer, wine, and liquor, along with other socioeconomic variables.

bmn2016

Format

A data frame with 1,722 observations and multiple variables related to liquor sales and policies.

id

Unique identifier for each observation

state

State name

year

Year of observation

beeroffweekday

Beer sales off-premises on weekdays

beerofffri

Beer sales off-premises on Fridays

beeroffsat

Beer sales off-premises on Saturdays

beeroffsun

Beer sales off-premises on Sundays

wineoffweekday

Wine sales off-premises on weekdays

wineofffri

Wine sales off-premises on Fridays

wineoffsat

Wine sales off-premises on Saturdays

wineoffsun

Wine sales off-premises on Sundays

liqoffweekday

Liquor sales off-premises on weekdays

liqofffri

Liquor sales off-premises on Fridays

liqoffsat

Liquor sales off-premises on Saturdays

liqoffsun

Liquor sales off-premises on Sundays

beeronweekday

Beer sales on-premises on weekdays

beeronfri

Beer sales on-premises on Fridays

beeronsat

Beer sales on-premises on Saturdays

beeronsun

Beer sales on-premises on Sundays

wineonweekday

Wine sales on-premises on weekdays

wineonfri

Wine sales on-premises on Fridays

wineonsat

Wine sales on-premises on Saturdays

wineonsun

Wine sales on-premises on Sundays

liqonweekday

Liquor sales on-premises on weekdays

liqonfri

Liquor sales on-premises on Fridays

liqonsat

Liquor sales on-premises on Saturdays

liqonsun

Liquor sales on-premises on Sundays

beerethvol

Beer ethanol volume

wineethvol

Wine ethanol volume

liqethvol

Liquor ethanol volume

beerage18

Legal beer drinking age 18

beerage19

Legal beer drinking age 19

beerage20

Legal beer drinking age 20

beerage21

Legal beer drinking age 21

wineage18

Legal wine drinking age 18

wineage19

Legal wine drinking age 19

wineage20

Legal wine drinking age 20

wineage21

Legal wine drinking age 21

liqage18

Legal liquor drinking age 18

liqage19

Legal liquor drinking age 19

liqage20

Legal liquor drinking age 20

liqage21

Legal liquor drinking age 21

unempw

Unemployment rate

popover18

Population over 18

beereth

Beer ethanol content

wineeth

Wine ethanol content

liqeth

Liquor ethanol content

logliq

Log of liquor sales

logwine

Log of wine sales

logbeer

Log of beer sales

expand

Indicator for expansion of sales

restrict

Indicator for restriction of sales

liqoffsunp

Liquor off-premises sales on Sundays (percentage)

liqonsunp

Liquor on-premises sales on Sundays (percentage)

beeroffinflows

Beer off-premises inflows

wineoffinflows

Wine off-premises inflows

liqoffinflows

Liquor off-premises inflows

beeroninflows

Beer on-premises inflows

wineoninflows

Wine on-premises inflows

liqoninflows

Liquor on-premises inflows

liqonoutflows

Liquor on-premises outflows

liqoffoutflows

Liquor off-premises outflows

beeronoutflows

Beer on-premises outflows

beeroffoutflows

Beer off-premises outflows

wineonoutflows

Wine on-premises outflows

wineoffoutflows

Wine off-premises outflows

liqonnotsun

Liquor on-premises sales not on Sundays

liqoffnotsun

Liquor off-premises sales not on Sundays

everonchange

Indicator for any change in on-premises sales

everoffchange

Indicator for any change in off-premises sales

winetax

Wine tax

liquortax

Liquor tax

beertax

Beer tax

cigtax

Cigarette tax

barsmokingban

Ban on smoking in bars

restsmokingban

Ban on smoking in restaurants

Source

Bernheim, B. D., Meer, J., & Novarro, N. K. (2016). "Do Consumers Exploit Commitment Opportunities? Evidence from Natural Experiments Involving Liquor Consumption." American Economic Journal: Economic Policy, 8(1), 41-69. Data available at https://www.aeaweb.org/articles?id=10.1257/pol.20130351