Card (1995) Dataset on College Proximity and Return to Schooling

Education, earnings, family background, and regional variables from the 1976 cross-section of the National Longitudinal Survey (NLS) of young men.

card1995

Format

A data frame with 3,613 observations and 52 variables. Missing values are coded as NA.

id

Sequential ID, runs from 1 to 5225

nearc2

Grew up near a 2-year college

nearc4

Grew up near a 4-year college

nearc4a

Grew up near a 4-year public college

nearc4b

Grew up near a 4-year private college

ed76

Education level in 1976

ed66

Education level in 1966

age76

Age in 1976

daded

Father's education (missing values are set to average)

nodaded

Indicator if father's education is imputed

momed

Mother's education

nomomed

Indicator if mother's education is imputed

weight

NLS weight for 1976 cross-section

momdad14

Indicator if lived with both parents at age 14

sinmom14

Indicator if lived with single mother at age 14

step14

Indicator if lived with step-parent at age 14

reg661

Dummy for region 1 in 1966

reg662

Dummy for region 2 in 1966

reg663

Dummy for region 3 in 1966

reg664

Dummy for region 4 in 1966

reg665

Dummy for region 5 in 1966

reg666

Dummy for region 6 in 1966

reg667

Dummy for region 7 in 1966

reg668

Dummy for region 8 in 1966

reg669

Dummy for region 9 in 1966

south66

Indicator if lived in the South in 1966

work76

Indicator if worked in 1976

work78

Indicator if worked in 1978

lwage76

Log wage in 1976 (outliers trimmed)

lwage78

Log wage in 1978 (outliers trimmed)

famed

Combined education level of mother and father

black

Indicator if Black

smsa76r

Indicator if lived in SMSA in 1976

smsa78r

Indicator if lived in SMSA in 1978

reg76r

Indicator if lived in the South in 1976

reg78r

Indicator if lived in the South in 1978

reg80r

Indicator if lived in the South in 1980

smsa66r

Indicator if lived in SMSA in 1966

wage76

Raw wage in cents per hour in 1976

wage78

Raw wage in cents per hour in 1978

wage80

Raw wage in cents per hour in 1980

noint78

Indicator if non-interviewed in 1978

noint80

Indicator if non-interviewed in 1980

enroll76

Indicator if enrolled in 1976

enroll78

Indicator if enrolled in 1978

enroll80

Indicator if enrolled in 1980

kww

Knowledge of the World of Work (KWW) score

iq

Normed IQ score

marsta76

Married and with spouse present in 1976

marsta78

Married and with spouse present in 1978

marsta80

Married and with spouse present in 1980

libcrd14

Indicator if had a library card at home at age 14

Source

Card, D. (1995). "Using Geographic Variation in College Proximity to Estimate the Return to Schooling." In L.N. Christofides, E.K. Grant, and R. Swidinsky (Eds.), Aspects of Labor Market Behaviour: Essays in Honour of John Vanderkamp. Toronto: University of Toronto Press. Data available at http://davidcard.berkeley.edu/data_sets.html