Card and Krueger (1994) Dataset on Minimum Wages and Employment

Employment and wage characteristics in fast-food restaurants across New Jersey and Pennsylvania. The dataset includes information on store characteristics, wages, and employee counts, reorganized so that each row represents one observation per store across two survey waves.

ck1994

Format

A data frame with 820 observations and 26 variables, representing 410 stores surveyed twice.

store

Unique store ID

chain

Fast-food chain

co_owned

Company-owned

state

State

southj

Located in southern New Jersey

centralj

Located in central New Jersey

northj

Located in northern New Jersey

pa1

Located in the northeast suburbs of Philadelphia

pa2

Located in the outside northeast suburbs of Philadelphia

shore

Located on New Jersey shore

ncalls

Number of call-backs (0 if contacted on first call)

empft

Number of full-time employees

emppt

Number of part-time employees

nmgrs

Number of managers/assistant managers

wage_st

Starting wage (dollars per hour)

inctime

Months to usual first raise

firstinc

Usual amount of first raise (dollars per hour)

meals

Free/reduced price meal code

open

Hour of opening

hoursopen

Number of hours open per day

pricesoda

Price of medium soda, including tax

pricefry

Price of small fries, including tax

priceentree

Price of entrée, including tax

nregisters

Number of cash registers in store

nregisters11

Number of registers open at 11am

time

Survey wave indicator

Source

Card, D., & Krueger, A. B. (1994). "Minimum Wages and Employment: A Case Study of the Fast-Food Industry in New Jersey and Pennsylvania." American Economic Review. Data available at http://davidcard.berkeley.edu/data_sets.html