Current Population Survey (March 2009) Labor Force Characteristics Dataset

Labor force characteristics for full-time employed U.S. civilians. The sample includes individuals who worked at least 36 hours per week for at least 48 weeks in the past year, excluding military personnel. The dataset contains 50,742 observations and 12 variables.

Usage

cps09mar

Format

A data frame with 50,742 observations and 12 variables.

age

Age in years, capped at 85

female

Indicator variable for female

female_num

Female indicator as numeric variable

hisp

Indicator variable for Spanish, Hispanic, or Latino

hisp_num

Hispanic indicator as numeric variable

education

Educational attainment

education_num

Education level as numeric variable

earnings

Total annual wage and salary earnings

hours

Number of hours worked per week

week

Number of weeks worked per year

union

1 if a member of a labor union, 0 otherwise

uncov

1 if covered by a union or employee association contract, 0 otherwise

region

Region of residence

region_num

Region of residence as numeric variable

race

Race

race_num

Race as numeric variable

marital

Marital status, coded as follows

marital_num

Marital status as numeric variable

Source

Bureau of Labor Statistics, U.S. Census Bureau. March 2009 Current Population Survey. More information available at https://www.census.gov/cps and https://dataferrett.census.gov.

Examples

# Explore the categorical variables codification
levels(cps09mar$female)
#> [1] "No"  "Yes"
levels(cps09mar$hisp)
#> [1] "No"  "Yes"
levels(cps09mar$education)
#>  [1] "Less than 1st grade"                                          
#>  [2] "1st to 4th grade"                                             
#>  [3] "5th or 6th grade"                                             
#>  [4] "7th or 8th grade"                                             
#>  [5] "9th grade"                                                    
#>  [6] "10th grade"                                                   
#>  [7] "11th grade or 12th grade without diploma"                     
#>  [8] "High school graduate or equivalent"                           
#>  [9] "Some college, no degree"                                      
#> [10] "Associate degree (occupational/vocational programs included)" 
#> [11] "Bachelor's degree (BA, AB, BS)"                               
#> [12] "Master's degree (MA, MS, MEng, MEd, MSW, MBA)"                
#> [13] "Professional or Doctorate degree (MD, DDS, DVM, JD, PhD, EdD)"
levels(cps09mar$region)
#> [1] "Northeast" "Midwest"   "South"     "West"     
levels(cps09mar$race)
#>  [1] "White only"                               
#>  [2] "Black only"                               
#>  [3] "American Indian, Alaskan Native (AI) only"
#>  [4] "Asian only"                               
#>  [5] "Hawaiian/Pacific Islander (HP) only"      
#>  [6] "White-Black"                              
#>  [7] "White-AI"                                 
#>  [8] "White-Asian"                              
#>  [9] "White-HP"                                 
#> [10] "Black-AI"                                 
#> [11] "Black-Asian"                              
#> [12] "Black-HP"                                 
#> [13] "AI-Asian"                                 
#> [14] "Asian-HP"                                 
#> [15] "White-Black-AI"                           
#> [16] "White-Black-Asian"                        
#> [17] "White-AI-Asian"                           
#> [18] "White-Asian-HP"                           
#> [19] "White-Black-AI-Asian"                     
#> [20] "2 or 3 races"                             
#> [21] "4 or 5 races"                             
levels(cps09mar$marital)
#> [1] "Married - civilian spouse present"         
#> [2] "Married - Armed Forces spouse present"     
#> [3] "Married - spouse absent (except separated)"
#> [4] "Widowed"                                   
#> [5] "Divorced"                                  
#> [6] "Separated"                                 
#> [7] "Never married"