McCracken and Ng (2015) FRED-QD Quarterly Database for Macroeconomic Research

Quarterly collection of major U.S. macroeconomic variables maintained by the St. Louis Federal Reserve Bank. This dataset provides key economic indicators across sectors such as output, labor, housing, consumption, interest rates, money, productivity, prices, and exchange rates. Includes data from 1959 onward.

Usage

fred_qd

Format

A data frame with 236 observations and 249 variables:

gdpc1

Real Gross Domestic Product

pcecc96

Real Personal Consumption Expenditures

pcdgx

Real Personal Consumption Expenditures: Durable Goods

pcesvx

Real Personal Consumption Expenditures: Services excluding food and energy

pcndx

Real Personal Consumption Expenditures: Nondurable Goods

gpdic1

Gross Private Domestic Investment

fpix

Fixed Private Investment

y033rc1q027sbeax

Real Gross Private Domestic Investment: Residential

pnfix

Private Nonresidential Fixed Investment

prfix

Private Residential Fixed Investment

a014re1q156nbea

Government Consumption Expenditures

gcec1

Real Government Consumption Expenditures and Gross Investment

a823rl1q225sbea

Real Government Spending

fgrecptx

Federal Government Current Receipts

slcex

State and Local Government Current Expenditures

expgsc1

Exports of Goods and Services

impgsc1

Imports of Goods and Services

dpic96

Disposable Personal Income

outnfb

Nonfarm Business Sector: Output

outbs

Business Sector Output

outms

Manufacturing Sector Output

indpro

Industrial Production Index

ipfinal

Industrial Production: Final Products

ipcongd

Industrial Production: Consumer Goods

ipmat

Industrial Production: Materials

ipdmat

Industrial Production: Durable Materials

ipnmat

Industrial Production: Nondurable Materials

ipdcongd

Industrial Production: Durable Consumer Goods

ipb51110sq

Industrial Production: Utilities

ipncongd

Industrial Production: Nondurable Consumer Goods

ipbuseq

Industrial Production: Business Equipment

ipb51220sq

Industrial Production: Residential Utilities

tcu

Total Capacity Utilization

cumfns

Nonfarm Business Sector: Total Compensation

payems

Total Nonfarm Payroll Employment

uspriv

All Employees: Private Industries

manemp

All Employees: Manufacturing

srvprd

All Employees: Service-Providing Industries

usgood

All Employees: Goods-Producing Industries

dmanemp

All Employees: Durable Goods

ndmanemp

All Employees: Nondurable Goods

uscons

All Employees: Construction

usehs

All Employees: Education and Health Services

usfire

All Employees: Financial Activities

usinfo

All Employees: Information

uspbs

All Employees: Professional and Business Services

uslah

All Employees: Leisure and Hospitality

usserv

All Employees: Other Services

usmine

All Employees: Mining and Logging

ustpu

All Employees: Trade, Transportation, and Utilities

usgovt

All Employees: Government

ustrade

All Employees: Retail Trade

uswtrade

All Employees: Wholesale Trade

ces9091000001

All Employees: State Government Education

ces9092000001

All Employees: State Government Excluding Education

ces9093000001

All Employees: Local Government Education

ce16ov

Civilian Employment: 16 Years and Over

civpart

Civilian Labor Force Participation Rate

unrate

Civilian Unemployment Rate

unratestx

Unemployment Rate for State

unrateltx

Unemployment Rate for Less than High School

lns14000012

Unemployment Rate: Hispanic or Latino

lns14000025

Unemployment Rate: 25-54 Years

lns14000026

Unemployment Rate: 55 Years and Over

uemplt5

Civilians Unemployed for Less Than 5 Weeks

uemp5to14

Civilians Unemployed for 5 to 14 Weeks

uemp15t26

Civilians Unemployed for 15 to 26 Weeks

uemp27ov

Civilians Unemployed for 27 Weeks and Over

lns13023621

Unemployment Rate for Bachelor’s Degree or Higher

lns13023557

Unemployment Rate for Associate Degree Holders

lns13023705

Unemployment Rate for Master’s Degree Holders

lns13023569

Unemployment Rate for High School Graduates

lns12032194

Labor Force Participation Rate for Bachelor’s Degree

hoabs

Housing Starts: Buildings with 5 or More Units

hoams

Housing Starts: All Buildings

hoanbs

Nonfarm Business Sector: Hours Worked

awhman

Average Weekly Hours: Manufacturing

awhnonag

Average Weekly Hours: Nonagricultural Industries

awotman

Average Weekly Overtime Hours: Manufacturing

hwix

Help-Wanted Index

houst

Housing Starts: Total

houst5f

Housing Starts: 5-Unit Structures or More

permit

New Private Housing Permits Authorized

houstmw

Housing Starts: Midwest Region

houstne

Housing Starts: Northeast Region

housts

Housing Starts: South Region

houstw

Housing Starts: West Region

cmrmtsplx

Manufacturing and Trade Sales

rsafsx

Residential Fixed Assets: Structures

amdmnox

New Orders for Durable Goods

acognox

New Orders: Consumer Goods and Materials

amdmuox

Unfilled Orders: Durable Goods Industries

andenox

Nondefense Capital Goods Excluding Aircraft

invcqrmtspl

Inventories: Manufacturing and Trade Industries

pcectpi

Personal Consumption Expenditures Chain-Type Price Index

pcepilfe

PCE: Food and Energy

gdpctpi

Gross Domestic Product Chain-Type Price Index

gpdictpi

Gross Private Domestic Investment Chain-Type Price Index

ipdbs

Industrial Production: Business Supplies

dgdsrg3q086sbea

Durable Goods Expenditures

ddurrg3q086sbea

Durable Goods Consumption Expenditures

dserrg3q086sbea

Services Expenditures

dndgrg3q086sbea

Nondurable Goods Expenditures

dhcerg3q086sbea

Health Care Expenditures

dmotrg3q086sbea

Motor Vehicle Expenditures

dfdhrg3q086sbea

Food at Home Expenditures

dreqrg3q086sbea

Recreational Goods Expenditures

dodgrg3q086sbea

Other Durable Goods Expenditures

dfxarg3q086sbea

Furniture and Fixtures Expenditures

dclorg3q086sbea

Clothing and Footwear Expenditures

dgoerg3q086sbea

Goods Expenditures

dongrg3q086sbea

Nondurable Goods Consumption Expenditures

dhutrg3q086sbea

Household Utilities Expenditures

dhlcrg3q086sbea

Household Cleaning Expenditures

dtrsrg3q086sbea

Transportation Services Expenditures

drcarg3q086sbea

Recreational Services Expenditures

dfsarg3q086sbea

Food Services and Accommodations

difsrg3q086sbea

Insurance Services Expenditures

dotsrg3q086sbea

Other Services Expenditures

cpiaucsl

Consumer Price Index for All Urban Consumers: All Items

cpilfesl

Consumer Price Index for All Urban Consumers: All Items Less Food and Energy

wpsfd49207

Producer Price Index by Commodity for Crude Materials

ppiaco

Producer Price Index by Commodity

wpsfd49502

Producer Price Index by Commodity for Intermediate Materials

wpsfd4111

Producer Price Index by Commodity for Finished Goods

ppiidc

Producer Price Index: Industrial Commodities

wpsid61

Producer Price Index for Finished Goods

wpu0531

Producer Price Index for Gasoline

wpu0561

Producer Price Index for Fuel Oil

oilpricex

Crude Oil Prices

ahetpix

Average Hourly Earnings of Production Employees

ces2000000008x

Average Hourly Earnings: Construction

ces3000000008x

Average Hourly Earnings: Manufacturing

comprms

Compensation per Hour: Manufacturing

comprnfb

Compensation per Hour: Nonfarm Business

rcphbs

Real Compensation Per Hour

ophmfg

Output per Hour: Manufacturing

ophnfb

Output per Hour: Nonfarm Business

ophpbs

Output per Hour: Private Business

ulcbs

Unit Labor Cost: Business Sector

ulcmfg

Unit Labor Cost: Manufacturing

ulcnfb

Unit Labor Cost: Nonfarm Business

unlpnbs

Unit Nonlabor Payments

fedfunds

Effective Federal Funds Rate

tb3ms

3-Month Treasury Bill Secondary Market Rate

tb6ms

6-Month Treasury Bill Secondary Market Rate

gs1

1-Year Treasury Constant Maturity Rate

gs10

10-Year Treasury Constant Maturity Rate

mortgage30us

30-Year Fixed Rate Mortgage Average

aaa

Moody's AAA Corporate Bond Yield

baa

Moody's BAA Corporate Bond Yield

baa10ym

Baa Rated Corporate Bond Yield Spread Over 10-Year Treasury

mortg10yrx

10-Year Treasury Note Average Yield

tb6m3mx

6-Month Treasury Minus 3-Month Treasury Spread

gs1tb3mx

1-Year Treasury Minus 3-Month Treasury Spread

gs10tb3mx

10-Year Treasury Minus 3-Month Treasury Spread

cpf3mtb3mx

Commercial Paper Minus 3-Month Treasury Spread

ambslrealx

Adjusted Monetary Base

imfslx

Monetary Base, Not Seasonally Adjusted

m1realx

Real M1 Money Stock

m2realx

Real M2 Money Stock

mzmrealx

MZM Money Stock

busloansx

Commercial and Industrial Loans at All Commercial Banks

consumerx

Consumer Credit Outstanding

nonrevslx

Nonrevolving Consumer Credit

reallnx

Real Estate Loans

revolslx

Revolving Consumer Credit

totalslx

Total Loans and Leases

driwcil

Deposits and Reserves in Financial Institutions

tabshnox

Total Assets at Banks

tlbshnox

Total Liabilities at Banks

liabpix

Liabilities

tnwbshnox

Total Net Worth of Households and Nonprofits

nwpix

Net Worth at Market Value

taresax

Treasury Assets and Reserves

hnoremq027sx

Household Net Worth

tfaabshnox

Total Financial Assets

vxoclsx

CBOE S&P 100 Volatility Index

ussthpi

All-Transactions House Price Index for the United States

spcs10rsa

S&P/Case-Shiller 10-City Composite Home Price Index

spcs20rsa

S&P/Case-Shiller 20-City Composite Home Price Index

twexmmth

Trade Weighted U.S. Dollar Index: Major Currencies

exuseu

Exchange Rate: U.S. Dollar to Euro

exszusx

Switzerland / U.S. Foreign Exchange Rate

exjpusx

Japan / U.S. Foreign Exchange Rate

exusukx

U.K. / U.S. Foreign Exchange Rate

excausx

Canada / U.S. Foreign Exchange Rate

umcsentx

University of Michigan: Consumer Sentiment Index

usepuindxm

U.S. Economic Policy Uncertainty Index

b020re1q156nbea

Government Current Expenditures

b021re1q156nbea

Net Government Saving

gfdegdq188s

Federal Debt: Total Public Debt

gfdebtnx

Gross Federal Debt as Percent of GDP

ipmansics

Industrial Production: Manufacturing

ipb51222s

Industrial Production: Fuels

ipfuels

Industrial Production: Fuels

uempmean

Average Duration of Unemployment

ces0600000007

Average Weekly Hours: Goods-Producing Industries

totresns

Total Reserves of Depository Institutions

nonborres

Nonborrowed Reserves of Depository Institutions

gs5

5-Year Treasury Constant Maturity Rate

tb3smffm

3-Month Treasury Minus Fed Funds Rate Spread

t5yffm

5-Year Treasury Minus Fed Funds Rate Spread

aaaffm

Moody's AAA Corporate Bond Minus Fed Funds Rate Spread

wpsid62

Producer Price Index by Commodity for Finished Goods: Consumer

ppicmm

Producer Price Index: Commodities

cpiappsl

Consumer Price Index for All Urban Consumers

cpitrnsl

CPI: Transportation

cpimedsl

CPI: Medical Care

cusr0000sac

CPI for All Urban Consumers: Commodities

cusr0000sad

CPI for All Urban Consumers: Durables

cusr0000sas

CPI for All Urban Consumers: Services

cpiulfsl

CPI: All Items Less Food and Energy

cusr0000sa0l2

CPI: Energy

cusr0000sa0l5

CPI: Food

ces0600000008

Average Hourly Earnings: Goods-Producing Industries

dtcolnvhfnm

Consumer Motor Vehicle Loans Outstanding

dtcthfnm

Total Consumer Loans Outstanding

invest

Investment in Bank Credit

hwiuratiox

Help-Wanted Index to Unemployed Ratio

claimsx

Initial Claims for Unemployment Insurance

businvx

Total Business Inventories

isratiox

Inventory-to-Sales Ratio

conspix

Consumer Credit to Personal Income Ratio

cp3m

3-Month Commercial Paper Rate

compapff

Commercial Paper Minus Fed Funds Rate Spread

permitne

New Private Housing Permits Authorized: Northeast

permitmw

New Private Housing Permits Authorized: Midwest

permits

New Private Housing Permits Authorized

permitw

New Private Housing Permits Authorized: West

nikkei225

Nikkei 225 Stock Index

nasdaqcom

NASDAQ Composite Index

cusr0000sehc

CPI for All Urban Consumers: Education

tlbsnncbx

Total Liabilities for Nonfinancial Corporate Business

tlbsnncbbdix

Total Liabilities for Nonfinancial Corporate Business, Billion Dollars

ttaabsnncbx

Total Assets for Nonfinancial Corporate Business

tnwmvbsnncbx

Total Net Worth for Nonfinancial Corporate Business

tnwmvbsnncbbdix

Total Net Worth for Nonfinancial Corporate Business, Billion Dollars

tlbsnnbx

Total Liabilities for Nonfinancial Noncorporate Business

tlbsnnbbdix

Total Liabilities for Nonfinancial Noncorporate Business, Billion Dollars

tabsnnbx

Total Assets for Nonfinancial Noncorporate Business

tnwbsnnbx

Total Net Worth for Nonfinancial Noncorporate Business

tnwbsnnbbdix

Total Net Worth for Nonfinancial Noncorporate Business, Billion Dollars

cncfx

Real Personal Consumption Expenditures: Non-Durables and Durables

sp500

S&P 500 Index

spindust

S&P Industrial Index

spdivyield

S&P Dividend Yield

spperatio

S&P Price-Earnings Ratio

time

Time Period

Source

McCracken, M. W., & Ng, S. "FRED-QD: A Quarterly Database for Macroeconomic Research." Data available at https://research.stlouisfed.org/econ/mccracken/fred-databases/

Details

Most of the variables are provided in the most basic format, so transformations are required for typical uses (e.g., for growth rates).