Nerlove (1963) Electricity Supply Dataset

Costs and production for 145 American electric companies in 1955.

Usage

nerlove1963

Format

A data frame with 145 observations on the following variables:

cost

Total Cost in million dollars

output

Electricity output in billion kilowatt-hours (kWh)

plabor

Unit price of labor, represented as wage rate in dollars per hour

pcapital

Unit price of capital, given as an index

pfuel

Unit price of fuel in cents per million BTU (British Thermal Units)

Source

Nerlove, Marc. "Returns to Scale in Electricity Supply" (1963), Measurement in Economics, C. Christ et al., eds.