Papageorgiou, Saam and Schulte (2017) Substitution between Clean and Dirty Energy Inputs Dataset

Electricity production, capacity, and fuel input across clean and dirty energy technologies for 26 countries over the period 1995-2009.

Usage

pss2017

Format

A data frame with observations across multiple countries and years, with the following variables:

country

Country name

year

Year (1995-2009)

eg_total

Net electricity production (GWh) for all plants

ec_total

Net electrical capacity of all plants (MW)

ec_c

Net installed clean generation capacity (MW)

ec_d

Net installed dirty generation capacity (MW)

fu_c

Fuel input for clean electricity-generating technologies (TJ)

fu_d

Fuel input for dirty electricity-generating technologies (TJ)

fu_total

Total fuel input for electricity-generating technologies (TJ)

ec_c_alt

Alternative capital proxy: Capital stock associated with clean technologies (EIA-based)

ec_d_alt

Alternative capital proxy: Capital stock associated with dirty technologies (EIA-based)

Source

Chris Papageorgiou, Marianne Saam, and Patrick Schulte (2017). Substitution between clean and dirty energy inputs: A macroeconomic perspective. The Review of Economics and Statistics, 281-290. Available from Harvard Dataverse: https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/67QCH2

Details

Data extracted from the file "electricity_sector.dta" and sourced from the REStat journal website and Dataverse repository.