Angrist and Krueger (1991) Dataset on Compulsory School Attendance

Schooling and earnings based on a subsample of men born 1930-1939 from the 1980 census extract. The dataset contains variables related to age, education, earnings, marital status, and geographic information.

Usage

ak1991

Format

A data frame with 329,509 observations across 10 columns.

ageq

Age, including quarter of birth

edu

Highest grade completed

logwage

Log weekly earnings (in dollars)

married

Married and with spouse present

qob

Quarter of birth

black

Black race

region

Region of residence

smsa

Works in a standard metropolitan statistical area (SMSA)

yob

Year of birth

state

State of birth

Source

Angrist, J., & Krueger, A. (1991). "Does Compulsory School Attendance Affect Schooling and Earnings?" The Quarterly Journal of Economics. Data available at http://economics.mit.edu/faculty/angrist/data1/data/angkru1991