Angrist and Lavy (1999) Dataset on Class Size and Scholastic Achievement

Class sizes, scholastic achievement scores, and various classroom demographics for 4th and 5th grades. The dataset is derived from the merged files final4.dta and final5.dta with data cleaning adjustments to remove recording errors, missing observations, class sizes above 44, and enrollments below 6.

Usage

al1999

Format

A data frame with 4,067 classrooms and 31 variables.

schlcode

School code

enrollment

School enrollment (number of students in grade)

enrollment_boys

Number of boys in grade

enrollment_girls

Number of girls in grade

c_num4rd

Number of 4th grade classes in school

c_type

Number of special education classes

flgrm4

Percentage failed in grammar 4th grade in 1991

mrkgrm4

Mean mark in grammar 4th grade in 1991

ngrm4

Number of students in grammar 4th grade in 1991

flmth4

Percentage failed in math 4th grade in 1991

mrkmth4

Mean mark in math 4th grade in 1991

nmth4

Number of students in math 4th grade in 1991

towncode

Town code

popcode

Population code

grade

Class grade (4 or 5)

classid

Class sequence number

classize

Number of students in class

mathsize

Number of students tested in mathematics

avgmath

Mathematics score

passmath

Pass mathematics test indicator

verbsize

Number of students tested in reading

avgverb

Grammar score

passverb

Pass grammar test indicator

disadvantaged

School socioeconomic status: percent disadvantaged

c_num5rd

Number of 5th grade classes in school

flgrm5

Percentage failed in grammar 5th grade in 1991

mrkgrm5

Mean mark in grammar 5th grade in 1991

ngrm5

Number of students in grammar 5th grade in 1991

flmth5

Percentage failed in math 5th grade in 1991

mrkmth5

Mean mark in math 5th grade in 1991

nmth5

Number of students in math 5th grade in 1991

Source

Angrist, J., & Lavy, V. (1999). "Using Maimonides’ Rule to Estimate the Effect of Class Size on Scholastic Achievement." The Quarterly Journal of Economics. Data available at https://economics.mit.edu/faculty/angrist/data1/data/anglavy99