Card, Mas, and Rothstein (2008) Segregation Dynamics Dataset

Demographic, economic, and housing characteristics across U.S. census tracts from 1970 to 2000. The dataset includes variables related to population composition, income, housing, and transportation, reorganized from the original replication files.

Usage

cmr2008

Format

A data frame with 48,643 observations (census tracts) and 46 variables.

geo2000

Census tract code

unem_70

Unemployment rate for population 16+ years, 1970

unem_80

Unemployment rate for population 16+ years, 1980

unem_90

Unemployment rate for population 16+ years, 1990

unem_00

Unemployment rate for population 16+ years, 2000

msa

Metropolitan Statistical Area (MSA) code

city

Name of the Metropolitan Statistical Area (MSA)

pubtran_70

Fraction of workers using public transportation, 1970

pubtran_80

Fraction of workers using public transportation, 1980

pubtran_90

Fraction of workers using public transportation, 1990

pop_70

Population, 1970

pop_80

Population, 1980

pop_90

Population, 1990

pop_00

Population, 2000

fr_white_70

Fraction White, 1970

fr_white_80

Fraction White, 1980

fr_white_90

Fraction White, 1990

fr_white_00

Fraction White, 2000

fr_min_70

Fraction Minority, 1970

fr_min_80

Fraction Minority, 1980

fr_min_90

Fraction Minority, 1990

fr_min_00

Fraction Minority, 2000

chg_white_7080

Change in White share, 1970-1980

chg_white_8090

Change in White share, 1980-1990

chg_white_9000

Change in White share, 1990-2000

chg_min_7080

Change in Minority share, 1970-1980

chg_min_8090

Change in Minority share, 1980-1990

chg_min_9000

Change in Minority share, 1990-2000

chg_pop_7080

Change in population, 1970-1980

chg_pop_8090

Change in population, 1980-1990

chg_pop_9000

Change in population, 1990-2000

samp_70

Indicator for inclusion in 1970 sample

samp_80

Indicator for inclusion in 1980 sample

samp_90

Indicator for inclusion in 1990 sample

faminc_70

Log of mean family income, 1970

faminc_80

Log of mean family income, 1980

faminc_90

Log of mean family income, 1990

vac_70

Fraction of vacant homes, 1970

vac_80

Fraction of vacant homes, 1980

vac_90

Fraction of vacant homes, 1990

rent_70

Fraction of renter-occupied homes, 1970

rent_80

Fraction of renter-occupied homes, 1980

rent_90

Fraction of renter-occupied homes, 1990

oneunit_70

Fraction of single-unit homes, 1970

oneunit_80

Fraction of single-unit homes, 1980

oneunit_90

Fraction of single-unit homes, 1990

Source

Card, D., Mas, A., & Rothstein, J. (2008). "Tipping and the Dynamics of Segregation." Quarterly Journal of Economics. Data available at https://berkeley.app.box.com/v/tippingreplication