Ludwig and Miller (2007) Impact of Head Start Program on Children's Life Chances

Examines the effect of the Head Start program on child mortality and other socioeconomic indicators across U.S. counties. Includes variables on poverty rates, mortality rates by age and cause, and census data from 1960 for 2,783 U.S. counties.

lm2007

Format

A data frame with 2,783 rows and 17 variables:

state

State name.

povrate60

County poverty rate in 1960 (percent).

mort_age59_related_posths

Mortality rate for ages 5-9 from Head Start-related causes, 1973-1983.

mort_age59_injury_posths

Mortality rate for ages 5-9 from injuries, 1973-1983.

mort_age59_all_posths

Mortality rate for ages 5-9 from all causes, 1973-1983.

mort_age25plus_related_posths

Mortality rate for ages 25+ from Head Start-related causes, 1973-1983.

mort_age25plus_injuries_posths

Mortality rate for ages 25+ from injuries, 1973-1983.

mort_age59_related_prehs

Mortality rate for ages 5-9 from Head Start-related causes, 1959-1964.

census1960_pop

County population according to the 1960 Census.

census1960_pctsch1417

Percentage of population aged 14-17 attending school, 1960.

census1960_pctsch534

Percentage of population aged 5-34 attending school, 1960.

census1960_pctsch25plus

Percentage of population aged 25+ with a high school education or more, 1960.

census1960_pop1417

Population aged 14-17, 1960.

census1960_pop534

Population aged 5-34, 1960.

census1960_pop25plus

Population aged 25+, 1960.

census1960_pcturban

Percentage of urban population, 1960.

census1960_pctblack

Percentage of Black population, 1960.

Source

Ludwig, J., & Miller, D. L. (2007). "Does Head Start improve children's life chances? Evidence from a regression discontinuity design," Quarterly Journal of Economics 122(1), 159-1208.

Matias D. Cattaneo, Rocio Titiunik, and Gonzalo Vazquez-Bare (2017) "Comparing inference approaches for RD designs: A reexamination of the effect of head start on child mortality," Journal of Policy Analysis and Management, 36, 643-681.'

Details

Abridged version of "headstart.dta" downloaded from https://github.com/rdpackages-replication/CTV_2017_JPAM which is a subset of that used by Ludwig and Miller and was provided for replication associated with Cattaneo, Titiunik, and Vazquez-Bare. Counties were deleted for which povrate60 or mort_age59_related_posths were missing.